silencer: implement dedicated receiver, various tweaks
